requirements.txt
ultralytics==8.0.20
opencv-python-headless==4.8.1.78
streamlit==1.32.0
numpy
Pillow
